save point: conformance fixes, eager prop/position, alarms, dispatch order
VERIFIED against the working PalaceChat by cross-client comparison (same script, same room, diffed):
- WHONAME now resolves a user name from an id (was returning empty)
- 32-bit keyword arithmetic no longer wraps at 3147483647+1
- '!=' is now the exact inverse of '==' (case-insensitive equality)
- SETSPOTSTATELOCAL returns the state it set
- three edge-case commands answer where PalaceChat answers
Rooms 32001, 32003, 32010 now match the reference line for line (47/47 on 32001).

ALSO FIXED EARLIER AND VERIFIED: script effect ordering (a012e28); worn props visible in-handler; self position visible in-handler after SETPOS; alarms no longer cleared by a room re-description.

NOT FIXED / NOT VERIFIED - do not assume: the Colosseum still benefits the user. Headless chain run (chain-live3.log) shows 1 'You are in the Audience' message and 8 occurrences of CIHPLAY=2 (spectator demotion) against 57 of CIHPLAY=20. The dispatch-order change is real progress (the fighter placement SETPOS 322,267 / SETSPOTSTATE spot=6 state=1 now lands, which it never did before) but the demotion still fires over it. Leading untested hypothesis: alarm-scheduled actions decide at schedule time rather than fire time, carrying a stale decision forward.

Evidence and tooling: test rooms 32001-32010 on localhost (kept permanently); PalaceChat reference /tmp/work/ref/palacechat-ref.txt (251 values); latest chain run /tmp/work/chain-live3.log. Not published; auto-publish stays disabled.

+++ crates/iptscrae-palace/src/commands.rs
@@ -544,7 +544,7 @@ impl<H: PalaceHost> Host for PalaceCommands<H> {
                 Ok(Vec::new())
             }
             "USERNAME" => Ok(vec![Value::str(self.inner.get_self_user_name())]),
-            "USERID" | "WHOME" => Ok(vec![Value::Int(self.inner.get_self_user_id() as i32)]),
+            "USERID" | "WHOME" => Ok(vec![Value::Int(self.inner.get_self_user_id())]),
             other => Err(IptError::CommandUnavailable {
                 command: other.to_owned(),
             }),
@@ -554,7 +554,7 @@ impl<H: PalaceHost> Host for PalaceCommands<H> {
 
 fn int_arg(args: &[Value], index: usize) -> Result<i64> {
     match args.get(index) {
-        Some(Value::Int(n)) => Ok(i64::from(*n)),
+        Some(Value::Int(n)) => Ok(*n),
         Some(other) => Err(IptError::TypeMismatch {
             expected: "number operand",
             found: other.type_name(),
